chore: move some functions from game component into utils folder
